refactor: moved pipeline row to own file
